refactor: create initialState and utils modules for reducer
